backend/requirements.txt
absl-py==1.2.0
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.6
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
executing==0.8.3
fastapi==0.79.0
fastjsonschema==2.15.3
filelock==3.7.1
flatbuffers==1.12
fonttools==4.34.4
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
httptools==0.5.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.8.0
outcome==1.2.0
packaging==21.3
panda==0.3.1
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.9.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pydeck==0.7.1
Pygments==2.12.0
PyJWT==2.5.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-twitter==3.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
regex==2022.7.9
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.12.0
selenium==4.3.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.19.1
streamlit==1.11.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
tweepy==4.10.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.10
uvicorn==0.17.6
validators==0.20.0
virtualenv==20.15.1
watchdog==2.1.9
watchgod==0.8.2
wcwidth==0.2.5
webdriver-manager==3.8.2
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.1
widgetsnbextension==3.6.1
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.1
